projects/char3d_gen/pyproject.toml
[project]
name = "char3d-gen"
version = "0.1.0"
description = "Generate 3D character models from multi-view reference images"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision",
    "diffusers>=0.27",
    "transformers>=4.38",
    "accelerate",
    "trimesh[easy]",
    "fast-simplification",
    "Pillow>=10.0",
    "rembg",
    "onnxruntime",
    "open-clip-torch",
    "numpy",
    "hy3dgen",
]

[tool.uv]
# Force PyTorch wheels built for CUDA 12.x so nvcc 12.6 matches torch.version.cuda
# (otherwise pip may install cu130 PyTorch while your toolkit is 12.6 → custom_rasterizer build fails).

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
hy3dgen = { git = "https://github.com/Tencent/Hunyuan3D-2.git" }
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
